+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Unreleased
+==========
+
+- Change typing: ``VDOM`` is now the tuple at the top, ``VDOMNode`` is the factory and an individual triple
+
+
 0.4.0
 =====